GEM_BACKEND/requirements.txt
beautifulsoup4
bs4
CacheControl
cachetools
certifi
cffi
chardet
click
deta
dnspython
fastapi
firebase-admin
gcloud
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
h11
httplib2
httptools
idna
jws
msgpack
oauth2client
packaging
proto-plus
protobuf
pyasn1
pyasn1-modules
pycparser
pycryptodome
pydantic
pydub
pymongo
pyparsing
Pyrebase
python-dotenv
python-jwt
python-multipart
python-snappy
pytz
PyYAML
requests
requests-toolbelt
rsa
six
soupsieve
starlette
typing-extensions
uritemplate
urllib3
uvicorn
uvloop
watchgod
websockets
jwt
